챌린지 모델 생성 및 상세페이지 작업중

+++ lib/widgets/widget_progress_indicator.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class WProgressIndicator extends StatelessWidget {
-  final double percentage;
+  final int percentage;
 
   const WProgressIndicator({
     super.key,
